requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
black==19.10b0
click==7.1.2
colorlog==4.1.0
Django==3.0.8
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==2.0.1
pathspec==0.8.0
pycrypto==2.6.1
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
